revise connection check to use new endpoint
